servo_url

Struct ServoUrl

pub struct ServoUrl(pub(crate) Arc<Url>);

Tuple Fields
0: Arc<Url>
Implementations

impl ServoUrl

pub fn from_url(url: Url) -> Self

pub fn parse_with_base( base: Option<&Self>, input: &str, ) -> Result<Self, ParseError>

pub fn into_string(self) -> String

pub fn into_url(self) -> Url

pub fn get_arc(&self) -> Arc<Url>

pub fn as_url(&self) -> &Url

pub fn parse(input: &str) -> Result<Self, ParseError>

pub fn cannot_be_a_base(&self) -> bool

pub fn domain(&self) -> Option<&str>

pub fn fragment(&self) -> Option<&str>

pub fn path(&self) -> &str

pub fn origin(&self) -> ImmutableOrigin

pub fn scheme(&self) -> &str

pub fn is_secure_scheme(&self) -> bool

pub fn is_local_scheme(&self) -> bool

https://fetch.spec.whatwg.org/#local-scheme

pub fn is_special_scheme(&self) -> bool

https://url.spec.whatwg.org/#special-scheme

pub fn is_equal_excluding_fragments(&self, other: &ServoUrl) -> bool

https://url.spec.whatwg.org/#url-equivalence In the future this may be removed if the helper is added upstream in rust-url see https://github.com/servo/rust-url/issues/1063 for details

pub fn as_str(&self) -> &str

pub fn as_mut_url(&mut self) -> &mut Url

pub fn set_username(&mut self, user: &str) -> Result<(), UrlError>

pub fn set_ip_host(&mut self, addr: IpAddr) -> Result<(), UrlError>

pub fn set_password(&mut self, pass: Option<&str>) -> Result<(), UrlError>

pub fn set_fragment(&mut self, fragment: Option<&str>)

pub fn username(&self) -> &str

pub fn password(&self) -> Option<&str>

pub fn to_file_path(&self) -> Result<PathBuf, UrlError>

pub fn host(&self) -> Option<Host<&str>>

pub fn host_str(&self) -> Option<&str>

pub fn port(&self) -> Option<u16>

pub fn port_or_known_default(&self) -> Option<u16>

pub fn join(&self, input: &str) -> Result<ServoUrl, ParseError>

pub fn path_segments(&self) -> Option<Split<'_, char>>

pub fn query(&self) -> Option<&str>

pub fn from_file_path<P: AsRef<Path>>(path: P) -> Result<Self, UrlError>

pub fn debug_compact(&self) -> impl Display + '_

Return a non-standard shortened form of the URL. Mainly intended to be used for debug printing in a constrained space (e.g., thread names).

pub fn is_potentially_trustworthy(&self) -> bool

https://w3c.github.io/webappsec-secure-contexts/#potentially-trustworthy-url

pub fn matches_about_blank(&self) -> bool

https://html.spec.whatwg.org/multipage/#matches-about:blank